refactor(tests): deduplicate callback factory and use existing helpers
- Extract _make_light_callback to tests/helpers.py as shared
  make_light_callback, imported by test_callbacks and test_groups
- Replace raw dev_type.startswith("01"/"02") with is_light_device()
  and is_sensor_device() from helper.py
- Replace hardcoded light type list in test_callbacks with
  is_light_device()
- Fix deprecated asyncio.get_event_loop() -> get_running_loop() in
  tests/helpers.py
